finished styling, added logo
